refactor combat#target
