+++ docs/src/content/docs/api-legacy/Combat.mdx
@@ -18,9 +18,9 @@ Whether attacks are paused due to an active counter attack.
 Type: `boolean`
 
 ### target <Badge text="Getter" size="small" />	
-Returns information about the target.
+Gets the target of the player.
 
-Type: `null | TargetInfo`
+Type: `null | Monster | Avatar`
 
 ## Methods
