+++ docs/src/content/docs/api-legacy/Player.mdx
@@ -72,7 +72,7 @@ Type: `[number, number]`
 ### state <Badge text="Getter" size="small" />	
 The state of the player.
 
-Type: `PlayerState`
+Type: `EntityState`
 
 ## Methods
