+++ docs/src/content/docs/api-legacy/enums/EntityState.mdx
@@ -1,0 +1,18 @@
+---
+title: EntityState
+---
+
+## Members
+
+### Dead = `0`
+The entity is dead.
+
+---
+### Idle = `1`
+The entity is idle.
+
+---
+### InCombat = `2`
+The entity is in combat.
+
+---
