+++ docs/src/content/docs/api-legacy/typedefs/AvatarData.mdx
@@ -4,26 +4,17 @@ title: AvatarData
 
 | Name | Type | Description |
 |------|------|-------------|
-| afk | `boolean` |  |
 | auras | `Aura[]` |  |
-| bResting | `boolean` |  |
-| entID | `number` |  |
-| entType | `string` |  |
-| ID | `number` |  |
 | intHP | `number` |  |
 | intHPMax | `number` |  |
+| intState | `number` |  |
+| afk | `boolean` |  |
+| entID | `number` |  |
+| entType | `string` |  |
 | intLevel | `number` |  |
 | intMP | `number` |  |
 | intMPMax | `number` |  |
 | intSP | `number` |  |
-| intSPMax | `number` |  |
-| intState | `number` |  |
-| mvtd | `string` |  |
-| mvts | `string` |  |
-| px | `string` |  |
-| py | `string` |  |
-| showCloak | `boolean` |  |
-| showHelm | `boolean` |  |
 | strFrame | `string` |  |
 | strPad | `string` |  |
 | strUsername | `string` |  |
