+++ docs/src/content/docs/api-legacy/typedefs/BaseEntityData.mdx
@@ -1,0 +1,10 @@
+---
+title: BaseEntityData
+---
+
+| Name | Type | Description |
+|------|------|-------------|
+| auras | `Aura[]` |  |
+| intHP | `number` |  |
+| intHPMax | `number` |  |
+| intState | `number` |  |
