+++ docs/src/content/docs/api-legacy/typedefs/MonsterData.mdx
@@ -4,10 +4,11 @@ title: MonsterData
 
 | Name | Type | Description |
 |------|------|-------------|
-| iLvl | `number` |  |
+| auras | `Aura[]` |  |
 | intHP | `number` |  |
 | intHPMax | `number` |  |
 | intState | `number` |  |
+| iLvl | `number` |  |
 | MonID | `number` |  |
 | MonMapID | `number` |  |
 | sRace | `string` |  |
