+++ src/renderer/game/lib/models/Monster.ts
@@ -1,12 +1,16 @@
+import { BaseEntity, type BaseEntityData } from "./BaseEntity";
+
 /**
  * Represents a monster.
  */
-export class Monster {
+export class Monster extends BaseEntity {
   public constructor(
     /**
      * Data about this monster.
-     */ public data: MonsterData,
-  ) {}
+     */ public override data: MonsterData,
+  ) {
+    super(data);
+  }
 
   /**
    * The monMapID of the monster.
@@ -30,13 +34,6 @@ export class Monster {
   }
 
   /**
-   * The state of the monster.
-   */
-  public get state(): number {
-    return this.data.intState;
-  }
-
-  /**
    * The race of the monster.
    */
   public get race(): string {
@@ -49,36 +46,12 @@ export class Monster {
   public get name(): string {
     return this.data.strMonName;
   }
-
-  /**
-   * The monster's current HP.
-   */
-  public get hp(): number {
-    return this.data.intHP;
-  }
-
-  /**
-   * The monster's max HP.
-   */
-  public get maxHp(): number {
-    return this.data.intHPMax;
-  }
-
-  /**
-   * Whether the monster is alive.
-   */
-  public get alive(): boolean {
-    return this.hp > 0;
-  }
 }
 
-export type MonsterData = {
+export type MonsterData = BaseEntityData & {
   MonID: number;
   MonMapID: number;
   iLvl: number;
-  intHP: number;
-  intHPMax: number;
-  intState: number;
   sRace: string;
   strFrame: string;
   strMonName: string;
